### 2019-12-12 07fa9fd
Update missing DOIs from bibtex
