+++ paper/paper.bib
@@ -19,7 +19,8 @@
 @article{shahri2019propheno,
   title={ProPheno 1.0: An Online Dataset for Accelerating the Complete Characterization of the Human Protein-phenotype Landscape in Biomedical Literature},
   author={Shahri, Morteza Pourreza and Kahanda, Indika},
-  year={2019}
+  year={2019},
+  doi={10.7287/peerj.preprints.27479v2}
 }
 
 @inproceedings{galea2018sub,
@@ -27,7 +28,8 @@
   author={Galea, Dieter and Laponogov, Ivan and Veselkov, Kirill},
   booktitle={Proceedings of the BioNLP 2018 workshop},
   pages={56--66},
-  year={2018}
+  year={2018},
+  doi={10.18653/v1/w18-2307}
 }
 
 @article{abdeddaim2018mesh,
@@ -52,7 +54,8 @@
   number={5},
   pages={e0198030},
   year={2018},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pone.0198030}
 }
 
 @article{nikolov2018data,
@@ -77,5 +80,6 @@
   booktitle={2019 IEEE 21st International Conference on High Performance Computing and Communications; IEEE 17th International Conference on Smart City; IEEE 5th International Conference on Data Science and Systems (HPCC/SmartCity/DSS)},
   pages={1580--1586},
   year={2019},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/hpcc/smartcity/dss.2019.00217}
 }
