### 2019-12-12 #1951 @kyleniemeyer
Hi @titipata, it looks like your article is missing DOIs for a number of the references; could you add those? Please review all the references (including their formatting), in case others are missing DOIs as well.
