### 2019-12-12 #1951 @titipata
Hi @kyleniemeyer, definitely, thank you! I added the missing DOIs and will check the other references.
