+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={429},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/s12859-019-2958-3}
 }
 
 @article{miller2017automated,
