+++ paper/paper.bib
@@ -33,11 +33,14 @@
   doi={10.18653/v1/w18-2307}
 }
 
-@article{abdeddaim2018mesh,
-  title={The MeSH-gram Neural Network Model: Extending Word Embedding Vectors with MeSH Concepts for UMLS Semantic Similarity and Relatedness in the Biomedical Domain},
-  author={Abdedda{\"\i}m, Sa{\"\i}d and Vimard, Sylvestre and Soualmia, Lina Fatima},
-  journal={arXiv preprint arXiv:1812.02309},
-  year={2018}
+@article{abdeddaim2019mesh,
+  title={The MeSH-Gram Neural Network Model: Extending Word Embedding Vectors with MeSH Concepts for Semantic Similarity.},
+  author={Abdeddaim, Sa{\"\i}d and Vimard, Sylvestre and Soualmia, Lina F},
+  journal={Studies in health technology and informatics},
+  volume={264},
+  pages={5--9},
+  year={2019},
+  doi={10.3233/shti190172}
 }
 
 @article{achakulvisut2019claim,
