### 2019-12-13 2cca123
Update issue templates
Add bug report and feature request template
