+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,26 @@
+---
+name: Bug report
+about: Create a report to help us improve Pubmed Parser
+title: Bug report
+labels: bug
+assignees: ''
+
+---
+
+**Describe the bug**
+Please describe the unexpected outcome from Pubmed Parser here
+
+**To Reproduce**
+Please provide code to reproduce an issue and how to download the dataset or file
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Dependencies**
+ - Please provide OS and dependencies you use for Pubmed Parser (e.g. `lxml`, `unidecode`, `requests`)
+
+**Additional context**
+Add any other context about the problem here.
