### 2019-12-13 b671a3d
Update documentation
