+++ docs/convert_readme.py
@@ -1,14 +1,17 @@
 #!/usr/bin/env python
+import os
+
 
 if __name__ == '__main__':
 
-    index_file = list()
-    index_file.extend(['---\n',
-                       'layout: default\n',
-                       '---\n',
-                       '\n'])
+    index_file = [
+        '---\n',
+        'layout: default\n',
+        '---\n',
+        '\n'
+    ]
 
-    with open('../README.md') as f:
+    with open(os.path.join('..', 'README.md')) as f:
         lines = f.readlines()
         for l in lines:
             index_file.append(l)
