+++ docs/index.md
@@ -2,260 +2,245 @@
 layout: default
 ---
 
-# Parser for Pubmed Open-Access XML Subset and MEDLINE XML Dataset
+# Pubmed Parser: A Python Parser for PubMed Open-Access XML Subset and MEDLINE XML Dataset
 
-[![Join the chat at https://gitter.im/titipata/pubmed_parser](https://badges.gitter.im/Join%20Chat.svg)](https://gitter.im/titipata/pubmed_parser?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge)
 [![License](https://img.shields.io/badge/License-MIT-blue.svg)](https://github.com/titipata/pubmed_parser/blob/master/LICENSE) [![DOI](https://zenodo.org/badge/31697087.svg)](https://zenodo.org/badge/latestdoi/31697087)
 
-Python parser for [PubMed Open-Access (OA) subset](http://www.ncbi.nlm.nih.gov/pmc/tools/ftp/)
- and [MEDLINE XML](https://www.nlm.nih.gov/bsd/licensee/) repository. See
+Pubmed Parser is a Python library for parsing [PubMed Open-Access (OA) subset](http://www.ncbi.nlm.nih.gov/pmc/tools/ftp/)
+ and [MEDLINE XML](https://www.nlm.nih.gov/bsd/licensee/) repository. It parses a dataset to dictionary which can be easily used for research such as text mining or natural language processing pipeline. See
  [wiki page](https://github.com/titipata/pubmed_parser/wiki) on how to download and
  process dataset using the repository.
 
-**note**
-- `path` provided to function can be path to compressed or uncompressed xml file.
-We provide example files in [`data`](data/) folder.
-- for website parser, you should scrape with pause. Please see
-[copyright notice](https://www.ncbi.nlm.nih.gov/pmc/about/copyright/#copy-PMC) because your IP
-can get blocked if you try to download in bulk.
+## Available Parsers
 
-## Parsers available
+* `path` provided to a function can be path to a compressed or uncompressed XML file. We provide example files in [ `data` ](data/) folder.
+* for website parser, you should scrape with pause. Please see [copyright notice](https://www.ncbi.nlm.nih.gov/pmc/about/copyright/#copy-PMC) because your IP can get blocked if you try to download in bulk.
 
-#### Parse Pubmed OA XML information
+Below, we list available parsers from `pubmed_parser`.
 
-We created a simple parser for PubMed Open Access Subset where you can give
-an XML path or string to the function called `parse_pubmed_xml` which will return
-a dictionary with the following information:
+  * [Parse PubMed OA XML information](#parse-pubmed-oa-xml-information)
+  * [Parse PubMed OA citation references](#parse-pubmed-oa-citation-references)
+  * [Parse PubMed OA images and captions](#parse-pubmed-oa-images-and-captions)
+  * [Parse PubMed OA Paragraph](#parse-pubmed-oa-paragraph)
+  * [Parse PubMed OA Table [WIP]](#parse-pubmed-oa-table-wip)
+  * [Parse MEDLINE XML](#parse-medline-xml)
+  * [Parse MEDLINE Grant ID](#parse-medline-grant-id)
+  * [Parse MEDLINE XML from eutils website](#parse-medline-xml-from-eutils-website)
+  * [Parse MEDLINE XML citations from website](#parse-medline-xml-citations-from-website)
+  * [Parse Outgoing XML citations from website](#parse-outgoing-xml-citations-from-website)
 
- - `full_title`: article's title
- - `abstract`: abstract
- - `journal`: Journal name
- - `pmid`: Pubmed ID
- - `pmc`: Pubmed Central ID
- - `doi`: DOI of the article
[…]
