### 2019-12-19 #1951 @majensen
Hello @timClicks -- would you be willing to review this artice (#1951) for [JOSS](https://joss.theoj.org)? Thanks for your consideration! --@majensen
