### 2019-12-19 #1951 @majensen
Hey @tleonardi, would you have time to review this article (#1951) for [JOSS](https://joss.theoj.org)? Prego! --@majensen
