### 2019-12-21 #1979 @whedon
[REVIEW]: Pubmed Parser: A Python Parser for PubMed Open-Access XML Subset and MEDLINE XML Dataset XML Dataset
**Submitting author:** @titipata (<a href="http://orcid.org/0000-0002-2124-2979">Titipat Achakulvisut</a>)
**Repository:** <a href="https://github.com/titipata/pubmed_parser" target ="_blank">https://github.com/titipata/pubmed_parser</a>
**Version:** v0.2.3
