### 2020-01-20 1477a0a
Solves SyntaxWarning raised by python 3.8. Issue #75
