### 2020-01-20 950eed2
Syntax/style changes
