+++ paper/paper.md
@@ -30,9 +30,9 @@ affiliations:
 
 # Summary
 
-Biomedical publications are increasing exponentially every year. If we had the ability to access, manipulate, and link this information, we could extract knowledge that is perhaps hidden within the figures, text, and citations. In particular, the repositories made available by [PubMed](https://pubmed.ncbi.nlm.nih.gov/) and [MEDLINE](https://www.nlm.nih.gov/bsd/medline.html) database enable these kinds of applications at an unprecedented level. Examples of applications that can be built from this dataset range from predicting novel drug-drug interactions, classifying biomedical text data, searching specific oncological profiles, disambiguating author names, or automatically learning a biomedical ontology. Here, we propose Pubmed Parser (`pubmed_parser`), a software to mine Pubmed and MEDLINE efficiently. Pubmed Parser is built on top of Python and can therefore be integrated into a myriad of tools for machine learning such as `scikit-learn` and deep learning such as `tensorflow` and `pytorch`.
+The number of biomedical publications is increasing exponentially every year. If we had the ability to access, manipulate, and link this information, we could extract knowledge that is perhaps hidden within the figures, text, and citations. In particular, the repositories made available by the [PubMed](https://pubmed.ncbi.nlm.nih.gov/) and [MEDLINE](https://www.nlm.nih.gov/bsd/medline.html) databases enable these kinds of applications at an unprecedented level. Examples of applications that can be built from this dataset range from predicting novel drug-drug interactions, classifying biomedical text data, searching specific oncological profiles, disambiguating author names, or automatically learning a biomedical ontology. Here, we propose Pubmed Parser (`pubmed_parser`), a software to mine Pubmed and MEDLINE efficiently. Pubmed Parser is built on top of Python and can therefore be integrated into a myriad of tools for machine learning such as `scikit-learn` and deep learning such as `tensorflow` and `pytorch`.
 
-Pubmed Parser has a capability to parse multiple refined information into structured dataset that former developed libraries such as [`medic`](https://github.com/fnl/medic) or [`MEDLINEXMLToJSON`](https://github.com/ldbib/MEDLINEXMLToJSON) do not have. It can parse multiple PubMed and MEDLINE data derivatives including articles' information, authors and affiliations, references, figure captions, paragraphs, and more. A core part of the package, parsing XML and HTML with Pubmed Parser is extremely fast, producing either dictionaries or JSON files that can easily be part of downstream pipelines. Moreover, the implemented functions can be scaled easily as part of other MapReduce-like infrastructure such as [PySpark](https://spark.apache.org/). This allows users to parse the most recently available corpus and customize the parsing to their needs. Below, we provide an example code to parse an XML file from [MEDLINE corpus](https://www.nlm.nih.gov/databases/download/data_distrib_main.html).
[…]
