### 2020-01-20 f88d422
Adding contributing guideline, update paper folder, update README
