+++ paper/paper.bib
@@ -70,9 +70,9 @@
 }
 
 @inproceedings{mesbah2018smartpub,
-  title={SmartPub: a platform for long-tail entity extraction from scientific publications},
+  title={SmartPub: A Platform for Long-Tail Entity Extraction from Scientific Publications},
   author={Mesbah, Sepideh and Bozzon, Alessandro and Lofi, Christoph and Houben, Geert-Jan},
-  booktitle={Companion Proceedings of the The Web Conference 2018},
+  booktitle={WWW},
   pages={191--194},
   year={2018},
   organization={International World Wide Web Conferences Steering Committee}
@@ -86,4 +86,11 @@
   year={2019},
   organization={IEEE},
   doi={10.1109/hpcc/smartcity/dss.2019.00217}
+}
+
+@article{abdeddaim2018mesh,
+  title={The MeSH-gram Neural Network Model: Extending Word Embedding Vectors with MeSH Concepts for UMLS Semantic Similarity and Relatedness in the Biomedical Domain},
+  author={Abdedda{\"\i}m, Sa{\"\i}d and Vimard, Sylvestre and Soualmia, Lina Fatima},
+  journal={arXiv preprint arXiv:1812.02309},
+  year={2018}
 }
