### 2020-01-20 901791d
Update documentation, contribution guideline, add contrib to README
