+++ CONTRIBUTING.md
@@ -2,7 +2,7 @@
 
 First of all, thanks for considering contributing to Pubmed Parser (`pubmed_parser`) 👍. We really appreciate your great help and support!
 
-`pubmed_parser` is an open source project, maintained by reseachers who want to make analysis of biomedical text and metadata easier.
+`pubmed_parser` is an open source Python project, maintained by reseachers who want to support and make analysis of biomedical text and metadata easy. We would like to accelerate how you get biomedical text data fast and, therefore, ease the effort in data cleaning process.
 
 ## How you can contribute?
 
@@ -10,11 +10,11 @@ There are several ways you can contribute to this project. Please see below
 
 ### Share the repository ❤️
 
-If you use `pubmed_parser` and think it is useful. Please feel free to share and tell us on Twitter.
+If you use `pubmed_parser` and think it is useful. Please feel free to share and tag us on Twitter [`@titipat_a`](https://twitter.com/titipat_a), [`@daniel_akuna`](https://twitter.com/daniel_akuna), and [`@KordingLab`](https://twitter.com/KordingLab).
 
 If you are using `pubmed_parser` on a paper you are writing, consider [citing it](https://github.com/titipata/pubmed_parser#citation).
 
-### Report a bug  🐛
+### Report a bug 🐛
 
 We use GitHub issues to track bugs, questions, and feature requests. Before reporting a bug, we suggest you to do [a quick search](https://github.com/titipata/pubmed_parser/issues) to see if your issue has already been reported. If your issue has already been reported, please comment on the existing issue.
 
@@ -27,12 +27,12 @@ Otherwise, open [a new GitHub issue](https://github.com/titipata/pubmed_parser/i
 
 For questions and feature requests, please frame it as narrow as possible so that we can discuss within the domain of your suggestions.
 
-### Propose an idea 💡
+### Propose a feature 💡
 
 Have an idea for a `pubmed_parser` feature? Take a look at the [repository](https://github.com/titipata/pubmed_parser) and [issue page](https://github.com/titipata/pubmed_parser/issues) to see if it isn't included or suggested yet. If not, suggest your idea as an issue on GitHub. While we can't promise to implement your idea, it helps to:
 
-* Explain in detail how it would work. - e.g. source of data and features that you want to parse
-* Keep the scope as narrow as possible.
+* Explain in detail how it would work. - provide source of data and features that you want us to add
+* Keep the scope as narrow as possible. - narrow scope makes it more feasible for us to complete your idea
 
 See below if you want to contribute code for your idea as well.
 
@@ -42,7 +42,7 @@ If you have been using `pubmed_parser` and see any errors that can be fixed, we 
 
 ### Ask a private question
 
-If you have a very specific research question which may not be directly related to the repository, we totally understand. You can email us for a private question. Although, we do **not** suggest you to email the maintainers directly.
+If you have a very specific research question which may not be directly related to the repository, you can email us for a private question. Although, we do **not** suggest you to email the maintainers directly. We encourage you to frame your question so that it can be asked directly on GitHub issue if possible.
 
 ## Development guidelines
 
@@ -58,4 +58,4 @@ We follow the [GitHub flow](https://guides.github.com/introduction/flow/) for de
 
 In your pull request, you can tag issues number relating to your pull request so that we can use it as a reference in the future. You can also give a short description on your pull request e.g. what is the improvement (code or documentation)? what error did you fix? etc.
 
-If you have see an inactivity on your pull request (it's likely that they might be really busy by that time), please re-tag core maintainers so that they can take a look closely to your pull requests again.
+If you have see an inactivity of your pull request, please re-tag core maintainers so that they can take a look closely to your pull requests again. Core maintainers might be really busy by the time you make the pull request.
