### 2020-01-20 ae58e0d
Minor update in README, contribution
