### 2020-01-21 8d3b739
copy edited paper.md
