### 2020-01-21 08f7454
some fixes to CONTRIBUTING.md
