### 2020-01-21 08c8867
some fixes to README.md
