### 2020-01-21 177869c
Add tests, and TravisCI integration
