+++ requirements.txt
@@ -1,3 +1,7 @@
 lxml
 unidecode
 requests
+six
+numpy
+pytest
+pytest-cov
