### 2020-01-21 c6ff97e
Code format using black, update part of example data, test script in README
