+++ pubmed_parser/pubmed_oa_parser.py
@@ -5,11 +5,11 @@ from .utils import *
 from unidecode import unidecode
 
 __all__ = [
-    'list_xml_path',
-    'parse_pubmed_xml',
-    'parse_pubmed_paragraph',
-    'parse_pubmed_references',
-    'parse_pubmed_caption'
+    "list_xml_path",
+    "parse_pubmed_xml",
+    "parse_pubmed_paragraph",
+    "parse_pubmed_references",
+    "parse_pubmed_caption",
 ]
 
 
@@ -25,8 +25,16 @@ def list_xml_path(path_dir):
     -------
     path_list: list, list of xml or nxml file from given path
     """
-    fullpath = [os.path.join(dp, f) for dp, dn, fn in os.walk(os.path.expanduser(path_dir)) for f in fn]
-    path_list = [folder for folder in fullpath if os.path.splitext(folder)[-1] in ('.nxml', '.xml')]
+    fullpath = [
+        os.path.join(dp, f)
+        for dp, dn, fn in os.walk(os.path.expanduser(path_dir))
+        for f in fn
+    ]
+    path_list = [
+        folder
+        for folder in fullpath
+        if os.path.splitext(folder)[-1] in (".nxml", ".xml")
+    ]
     return path_list
 
 
@@ -53,23 +61,18 @@ def parse_article_meta(tree):
     """
     Parse PMID, PMC and DOI from given article tree
     """
-    article_meta = tree.find('.//article-meta')
+    article_meta = tree.find(".//article-meta")
     pmid_node = article_meta.find('article-id[@pub-id-type="pmid"]')
     pmc_node = article_meta.find('article-id[@pub-id-type="pmc"]')
     pub_id_node = article_meta.find('article-id[@pub-id-type="publisher-id"]')
     doi_node = article_meta.find('article-id[@pub-id-type="doi"]')
 
-    pmid = pmid_node.text if pmid_node is not None else ''
-    pmc = pmc_node.text if pmc_node is not None else ''
-    pub_id = pub_id_node.text if pub_id_node is not None else ''
-    doi = doi_node.text if doi_node is not None else ''
+    pmid = pmid_node.text if pmid_node is not None else ""
+    pmc = pmc_node.text if pmc_node is not None else ""
+    pub_id = pub_id_node.text if pub_id_node is not None else ""
+    doi = doi_node.text if doi_node is not None else ""
 
-    dict_article_meta = {
-        'pmid': pmid,
-        'pmc': pmc,
-        'doi': doi,
-        'publisher_id': pub_id
-    }
+    dict_article_meta = {"pmid": pmid, "pmc": pmc, "doi": doi, "publisher_id": pub_id}
 
     return dict_article_meta
 
@@ -81,63 +84,63 @@ def parse_pubmed_xml(path, include_path=False, nxml=False):
     """
     tree = read_xml(path, nxml)
 
-    tree_title = tree.find('.//title-group/article-title')
+    tree_title = tree.find(".//title-group/article-title")
     if tree_title is not None:
         title = [t for t in tree_title.itertext()]
-        sub_title = tree.xpath('.//title-group/subtitle/text()')
+        sub_title = tree.xpath(".//title-group/subtitle/text()")
         title.extend(sub_title)
-        title = [t.replace('\n', ' ').replace('\t', ' ') for t in title]
-        full_title = ' '.join(title)
+        title = [t.replace("\n", " ").replace("\t", " ") for t in title]
+        full_title = " ".join(title)
     else:
-        full_title = ''
+        full_title = ""
 
     try:
         abstracts = list()
-        abstract_tree = tree.findall('.//abstract')
+        abstract_tree = tree.findall(".//abstract")
         for a in abstract_tree:
             for t in a.itertext():
[…]
