+++ pubmed_parser/__init__.py
@@ -3,16 +3,19 @@ Pubmed Parser: A Python Parser for PubMed Open-Access XML Subset and MEDLINE XML
 
 Author: Titipat Achakulvisut, Daniel E. Acuna
 """
-__version__ = '0.2.1'
-from .pubmed_oa_parser import list_xml_path, \
-                              parse_pubmed_xml, \
-                              parse_pubmed_references, \
-                              parse_pubmed_paragraph, \
-                              parse_pubmed_caption, \
-                              parse_pubmed_table
-from .medline_parser import parse_medline_xml, \
-                            parse_medline_grant_id
-from .pubmed_web_parser import parse_xml_web, \
-                               parse_citation_web, \
-                               parse_outgoing_citation_web
+__version__ = "0.2.1"
+from .pubmed_oa_parser import (
+    list_xml_path,
+    parse_pubmed_xml,
+    parse_pubmed_references,
+    parse_pubmed_paragraph,
+    parse_pubmed_caption,
+    parse_pubmed_table,
+)
+from .medline_parser import parse_medline_xml, parse_medline_grant_id
+from .pubmed_web_parser import (
+    parse_xml_web,
+    parse_citation_web,
+    parse_outgoing_citation_web,
+)
 from .utils import pretty_print
