+++ pubmed_parser/medline_parser.py
@@ -4,10 +4,7 @@ from itertools import chain
 from collections import defaultdict
 from pubmed_parser.utils import read_xml, stringify_children, month_or_day_formater
 
-__all__ = [
-    'parse_medline_xml',
-    'parse_medline_grant_id'
-]
+__all__ = ["parse_medline_xml", "parse_medline_grant_id"]
 
 
 def parse_pmid(pubmed_article):
@@ -23,17 +20,17 @@ def parse_pmid(pubmed_article):
     pmid: str
         String version of the PubMed ID
     """
-    medline = pubmed_article.find('MedlineCitation')
-    if medline.find('PMID') is not None:
-        pmid = medline.find('PMID').text
+    medline = pubmed_article.find("MedlineCitation")
+    if medline.find("PMID") is not None:
+        pmid = medline.find("PMID").text
         return pmid
     else:
-        article_ids = pubmed_article.find('PubmedData/ArticleIdList')
+        article_ids = pubmed_article.find("PubmedData/ArticleIdList")
         if article_ids is not None:
             pmid = article_ids.find('ArticleId[@IdType="pmid"]')
-            pmid = pmid.text.strip() if pmid is not None else ''
+            pmid = pmid.text.strip() if pmid is not None else ""
         else:
-            pmid = ''
+            pmid = ""
     return pmid
 
 
@@ -49,20 +46,24 @@ def parse_doi(pubmed_article):
     -------
     doi: str, DOI from a given lxml node
     """
-    medline = pubmed_article.find('MedlineCitation')
-    article = medline.find('Article')
-    elocation_ids = article.findall('ELocationID')
+    medline = pubmed_article.find("MedlineCitation")
+    article = medline.find("Article")
+    elocation_ids = article.findall("ELocationID")
 
     if len(elocation_ids) > 0:
         for e in elocation_ids:
-            doi = e.text.strip() or '' if e.attrib.get('EIdType', '') == 'doi' else ''
+            doi = e.text.strip() or "" if e.attrib.get("EIdType", "") == "doi" else ""
     else:
-        article_ids = pubmed_article.find('PubmedData/ArticleIdList')
+        article_ids = pubmed_article.find("PubmedData/ArticleIdList")
         if article_ids is not None:
             doi = article_ids.find('ArticleId[@IdType="doi"]')
-            doi = (doi.text.strip() if doi.text is not None else '') if doi is not None else ''
+            doi = (
+                (doi.text.strip() if doi.text is not None else "")
+                if doi is not None
+                else ""
+            )
         else:
-            doi = ''
+            doi = ""
     return doi
 
 
@@ -80,16 +81,17 @@ def parse_mesh_terms(medline):
         String of semi-colon spearated MeSH (Medical Subject Headings)
         terms contained in the document.
     """
-    if medline.find('MeshHeadingList') is not None:
-        mesh = medline.find('MeshHeadingList')
+    if medline.find("MeshHeadingList") is not None:
+        mesh = medline.find("MeshHeadingList")
         mesh_terms_list = [
-            m.find('DescriptorName').attrib.get('UI', '') +
-            ":" +
-            m.find('DescriptorName').text for m in mesh.getchildren()
+            m.find("DescriptorName").attrib.get("UI", "")
+            + ":"
+            + m.find("DescriptorName").text
+            for m in mesh.getchildren()
         ]
-        mesh_terms = '; '.join(mesh_terms_list)
+        mesh_terms = "; ".join(mesh_terms_list)
     else:
-        mesh_terms = ''
+        mesh_terms = ""
     return mesh_terms
 
 
@@ -107,14 +109,16 @@ def parse_publication_types(medline):
         String of semi-colon spearated publication types
     """
     publication_types = []
[…]
