### 2020-01-21 a5462ad
Tweak travis yml file and README
