+++ README.md
@@ -232,7 +232,7 @@ git clone https://github.com/titipata/pubmed_parser
 pip install ./pubmed_parser
 ```
 
-You can test your installation by running `py.test --cov=pubmed_parser tests/ --verbose`
+You can test your installation by running `pytest --cov=pubmed_parser tests/ --verbose`
 
 ## Example snippet to parse PubMed OA dataset
 
