+++ .travis.yml
@@ -1,25 +1,19 @@
 language: python
 env:
-  - PYTHON=3.5
-  - PYTHON=3.6
-  - PYTHON=3.7
-  - PYTHON=3.8
+  - 3.5
+  - 3.6
+  - 3.7
+  - 3.8
 
-miniconda:
-  - wget -q http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh
-  - chmod +x miniconda.sh
-  - ./miniconda.sh -b -p /home/travis/miniconda
-  - export PATH=/home/travis/miniconda/bin:$PATH
-  - conda update --yes conda
+before_install:
+  - python --version
+  - pip install -U pip
 
 install:
-  - conda create -n testenv --yes pip python=$PYTHON
-  - source activate testenv
-  - conda install --yes --file requirements.txt
   - python setup.py install
 
 script:
-  - py.test --cov=pubmed_parser tests/ --verbose
+  - pytest --cov=pubmed_parser tests/ --verbose
 
 after_success:
-  - bash <(curl -s https://codecov.io/bash)
+  - codecov
