### 2020-01-21 33fd8b3
Add TravisCI badge, update .travis.yml file
