+++ .travis.yml
@@ -1,9 +1,9 @@
 language: python
 env:
-  - 3.5
-  - 3.6
-  - 3.7
-  - 3.8
+  - "3.5"
+  - "3.6"
+  - "3.7"
+  - "3.8"
 
 before_install:
   - python --version
