### 2020-01-21 0ed9511
Update authorship, Travis yml file
