### 2020-01-22 43d324b
Update and format tests folder, update setup.py
