+++ setup.py
@@ -3,23 +3,32 @@ from setuptools import setup
 
 if __name__ == "__main__":
     setup(
-        name='pubmed_parser',
-        version='0.2.1',
-        description='A python parser for Pubmed Open-Access Subset and MEDLINE XML repository',
-        url='https://github.com/titipata/pubmed_parser',
-        download_url='https://github.com/titipata/pubmed_parser.git',
-        author='Titipat Achakulvisut',
-        author_email='my.titipat@gmail.com',
-        license='(c) 2015 - 2019 Titipat Achakulvisut, Daniel E. Acuna',
+        name="pubmed_parser",
+        version="0.2.1",
+        description="A python parser for Pubmed Open-Access Subset and MEDLINE XML repository",
+        url="https://github.com/titipata/pubmed_parser",
+        download_url="https://github.com/titipata/pubmed_parser.git",
+        author="Titipat Achakulvisut",
+        author_email="my.titipat@gmail.com",
+        license="MIT (c) 2015 - 2019 Titipat Achakulvisut, Daniel E. Acuna",
         install_requires=[
-            'lxml', 'unidecode', 'requests',
-            'six', 'numpy', 'pytest',
-            'pytest-cov'
+            "lxml",
+            "unidecode",
+            "requests",
+            "six",
+            "numpy",
+            "pytest",
+            "pytest-cov",
         ],
-        packages=['pubmed_parser'],
-        package_data={
-            'pubmed_parser.data': ['*.xml.gz', '*.nxml', '*.txt'],
-        },
+        packages=["pubmed_parser"],
+        package_data={"pubmed_parser.data": ["*.xml.gz", "*.nxml", "*.txt"],},
+        keywords=[
+            'Python',
+            'MEDLINE',
+            'PubMed',
+            'Biomedical corpus',
+            'Natural Language Processing'
+        ],
         classifiers=[
             "Intended Audience :: Science/Research",
             "Intended Audience :: Developers",
@@ -34,12 +43,12 @@ if __name__ == "__main__":
             "Programming Language :: Python :: 3.5",
             "Programming Language :: Python :: 3.6",
             "Programming Language :: Python :: 3.7",
-            "Programming Language :: Python :: 3.8"
+            "Programming Language :: Python :: 3.8",
         ],
-        platforms='any',
+        platforms="any",
         project_urls={
-            'Source': 'https://github.com/titipata/pubmed_parser',
-            'Documentation': 'http://titipata.github.io/pubmed_parser',
-            'Bug Reports': 'https://github.com/titipata/pubmed_parser/issues'
-        }
+            "Source": "https://github.com/titipata/pubmed_parser",
+            "Documentation": "http://titipata.github.io/pubmed_parser",
+            "Bug Reports": "https://github.com/titipata/pubmed_parser/issues",
+        },
     )
