+++ tests/test_medline_parser.py
@@ -20,3 +20,15 @@ def test_parse_medline_xml():
     assert parsed_medline[0]["title"][0:50] == expected_title
     assert parsed_medline[0]["abstract"][0:50] == expected_abstract
     assert parsed_medline[0]["pmid"] == "399296"
+
+
+def test_parse_medline_grant_id():
+    """
+    Test parsing grants from MEDLINE XML
+    """
+    grants = pp.parse_medline_grant_id(os.path.join("data", "pubmed20n0014.xml.gz"))
+    assert isinstance(grants, list)
+    assert isinstance(grants[0], dict)
+    assert grants[0]["pmid"] == "399300"
+    assert grants[0]["grant_id"] == "HL17731"
+    assert len(grants) == 484, "Expect number of grants in a given file to be 484"
