+++ docs/index.md
@@ -1,335 +1,0 @@
----
-layout: default
----
-
-# Pubmed Parser: A Python Parser for PubMed Open-Access XML Subset and MEDLINE XML Dataset
-
-[![License](https://img.shields.io/badge/License-MIT-blue.svg)](https://github.com/titipata/pubmed_parser/blob/master/LICENSE) [![DOI](https://zenodo.org/badge/31697087.svg)](https://zenodo.org/badge/latestdoi/31697087)
-
-Pubmed Parser is a Python library for parsing [PubMed Open-Access (OA) subset](http://www.ncbi.nlm.nih.gov/pmc/tools/ftp/)
- and [MEDLINE XML](https://www.nlm.nih.gov/bsd/licensee/) repository. It parses a dataset to dictionary which can be easily used for research such as text mining or natural language processing pipeline. See
- [wiki page](https://github.com/titipata/pubmed_parser/wiki) on how to download and
- process dataset using the repository.
-
-## Available Parsers
-
-* `path` provided to a function can be path to a compressed or uncompressed XML file. We provide example files in [ `data` ](data/) folder.
-* for website parser, you should scrape with pause. Please see [copyright notice](https://www.ncbi.nlm.nih.gov/pmc/about/copyright/#copy-PMC) because your IP can get blocked if you try to download in bulk.
-
-Below, we list available parsers from `pubmed_parser`.
-
-  * [Parse PubMed OA XML information](#parse-pubmed-oa-xml-information)
-  * [Parse PubMed OA citation references](#parse-pubmed-oa-citation-references)
-  * [Parse PubMed OA images and captions](#parse-pubmed-oa-images-and-captions)
-  * [Parse PubMed OA Paragraph](#parse-pubmed-oa-paragraph)
-  * [Parse PubMed OA Table [WIP]](#parse-pubmed-oa-table-wip)
-  * [Parse MEDLINE XML](#parse-medline-xml)
-  * [Parse MEDLINE Grant ID](#parse-medline-grant-id)
-  * [Parse MEDLINE XML from eutils website](#parse-medline-xml-from-eutils-website)
-  * [Parse MEDLINE XML citations from website](#parse-medline-xml-citations-from-website)
-  * [Parse Outgoing XML citations from website](#parse-outgoing-xml-citations-from-website)
-
-### Parse PubMed OA XML information
-
-We created a simple parser for PubMed Open Access Subset where you can give an XML path or string to the function called `parse_pubmed_xml` which will return a dictionary with the following information:
-
-* `full_title` : article's title
-* `abstract` : abstract
-* `journal` : Journal name
-* `pmid` : PubMed ID
-* `pmc` : PubMed Central ID
-* `doi` : DOI of the article
-* `publisher_id` : publisher ID
-* `author_list` : list of authors with affiliation keys in the following format
-
-``` python
- [['last_name_1', 'first_name_1', 'aff_key_1'],
-  ['last_name_1', 'first_name_1', 'aff_key_2'],
-  ['last_name_2', 'first_name_2', 'aff_key_1'], ...]
- ```
-
-* `affiliation_list` : list of affiliation keys and affiliation strings in the following format
-
-``` python
- [['aff_key_1', 'affiliation_1'],
-  ['aff_key_2', 'affiliation_2'], ...]
-```
-
-* `publication_year` : publication year
-* `subjects` : list of subjects listed in the article separated by semicolon. Sometimes, it only contains type of article, such as research article, review proceedings, etc.
-
-``` python
-import pubmed_parser as pp
-dict_out = pp.parse_pubmed_xml(path)
-```
-
-### Parse PubMed OA citation references
-
[…]
