+++ docs/README.md
@@ -1,10 +1,0 @@
-## Jekyll page for Pubmed Parser
-
-We can update the page by running `python convert_readme.py`. This will convert `README.md`
-to `index.md` (used for Jekyll page).
-
-We can serve the server locally to
-
-```bash
-jekyll serve
-```
