+++ docs/api.rst
@@ -1,0 +1,4 @@
+API Documentation
+=================
+
+.. currentmodule:: pubmed_parser
