+++ docs/api.rst
@@ -2,3 +2,25 @@ API Documentation
 =================
 
 .. currentmodule:: pubmed_parser
+
+Parse MEDLINE XML 
+-----------------
+
+.. autofunction:: parse_medline_xml
+.. autofunction:: parse_medline_grant_id
+
+Parse PubMed XML
+----------------
+
+.. autofunction:: parse_pubmed_xml
+.. autofunction:: parse_pubmed_references
+.. autofunction:: parse_pubmed_paragraph
+.. autofunction:: parse_pubmed_caption
+.. autofunction:: parse_pubmed_table
+
+Parse from Website
+------------------
+
+.. autofunction:: parse_xml_web
+.. autofunction:: parse_citation_web
+.. autofunction:: parse_outgoing_citation_web
