+++ docs/README.md
@@ -1,0 +1,6 @@
+# Pubmed Parser Documentation
+
+We use [`sphinx`](https://www.sphinx-doc.org/en/master/) to generate our documentation page.
+
+To build documentation page, run `make html`. All static files will be built in
+`_build/html/` where you can open them on the web browser.
