+++ _sources/index.rst.txt
@@ -1,0 +1,26 @@
+.. Pubmed Parser documentation master file, created by
+   sphinx-quickstart on Tue Jan 21 20:26:52 2020.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Pubmed Parser: A Python Parser for PubMed Open-Access XML Subset and MEDLINE XML Dataset
+========================================================================================
+
+
+Pubmed Parser is a Python library for parsing the `PubMed Open-Access (OA) subset <http://www.ncbi.nlm.nih.gov/pmc/tools/ftp//>`_
+and `MEDLINE XML <https://www.nlm.nih.gov/bsd/licensee//>`_ repositories. 
+It parses this information into a Python dictionary which can be easily used for research such in text mining and natural language processing pipelines. 
+See `wiki page <https://github.com/titipata/pubmed_parser/wiki/>`_ on how to download and process dataset using the repository.
+
+
+
+Contents
+========
+
+.. toctree::
+   :maxdepth: 1
+
+   install
+   api
+   resources
+   spark
