+++ .travis.yml
@@ -1,0 +1,22 @@
+language: python
+env:
+  - 3.5
+  - 3.6
+  - 3.7
+  - 3.8
+
+notifications:
+  email: false
+
+before_install:
+  - python --version
+  - pip install -U pip
+
+install:
+  - python setup.py install
+
+script:
+  - pytest --cov=pubmed_parser tests/ --verbose
+
+after_success:
+  - bash <(curl -s https://codecov.io/bash)
