### 2020-01-22 #1979 @titipata
Hi all,

Just a quick update here. I updated the repository according to @tleonardi comments. I also created a test folder and sphinx documentation on the repository which I will be updating in this coming week. We hope to make the documentation better in the coming week!
