### 2020-01-22 7307b1c
Major documentation updates
