### 2020-01-22 ce35d88
Minor update documentation
