+++ _sources/resources.rst.txt
@@ -10,10 +10,10 @@ Links to download Pubmed and MEDLINE dataset
 
 Here are links for downloading PubMed OA and MEDLINE data
 
-- PubMed Open-Access (OA) dataset is available at ``http://www.ncbi.nlm.nih.gov/pmc/tools/ftp/``. Here is the `FTP link <ftp://ftp.ncbi.nlm.nih.gov/pub/pmc//>`_ for downloading the bulk of dataset.
+- PubMed Open-Access (OA) dataset is available at ``http://www.ncbi.nlm.nih.gov/pmc/tools/ftp/``. Here is the `FTP link <ftp://ftp.ncbi.nlm.nih.gov/pub/pmc/>`_ for downloading the bulk of dataset. You can check `oa_bulk folder <ftp://ftp.ncbi.nlm.nih.gov/pub/pmc/oa_bulk//>`_ to see the full tar files.
 - the MEDLINE XMLs are available here ``ftp://ftp.nlm.nih.gov/nlmdata/.medleasebaseline/gz/``
 - the MEDLINE XMLs weekly updates are available here ``ftp://ftp.nlm.nih.gov/nlmdata/.medlease/gz/``
-- MEDLINE Document Type Definitions (DTDs) file is available at this `<link https://www.nlm.nih.gov/databases/dtd//>`_. We can use it to see available tags from a given MEDLINE XML.
+- MEDLINE Document Type Definitions (DTDs) file is available at this `link <https://www.nlm.nih.gov/databases/dtd/>`_. We can use it to see available tags from a given MEDLINE XML.
 
 
 Download PubMed OA figures
@@ -21,19 +21,22 @@ Download PubMed OA figures
 
 Here, we explain how to download PubMed OA figures corresponded to the parsed information
 
-- In `pubmed_parser`, you can use ``parse_pubmed_caption`` to parse figures (to be specific `figure_id`) and captions corresponding to a manuscript. 
-- To download the images corresponding to a given ``PMC`` or ``PMID``, you can download a CSV file from `<ftp://ftp.ncbi.nlm.nih.gov/pub/pmc/oa_file_list.csv <ftp://ftp.ncbi.nlm.nih.gov/pub/pmc/oa_file_list.csv/>`_ first. - The file will have columns ``PMID``, ``Accession ID`` (``PMC``), and ``File`` where it looks something like ``oa_package/08/e0/PMC13900.tar.gz``.
-- You can then download a tar file for a given ``PMID`` or ``PMC`` from `ftp://ftp.ncbi.nlm.nih.gov/pub/pmc/oa_package/08/e0/PMC13900.tar.gz <ftp://ftp.ncbi.nlm.nih.gov/pub/pmc/oa_package/08/e0/PMC13900.tar.gz/>`_. You can check out `ftp://ftp.ncbi.nlm.nih.gov/pub/pmc/oa_package/` to get all the access for tar files.
+- In ``pubmed_parser``, you can use ``parse_pubmed_caption`` to parse figures (to be specific ``figure_id``) and captions corresponding to a manuscript. 
+- To download the images corresponding to a given ``PMC`` or ``PMID``, you can download a CSV file from ``ftp://ftp.ncbi.nlm.nih.gov/pub/pmc/oa_file_list.csv`` first. - The file will have columns ``PMID``, ``Accession ID`` (``PMC``), and ``File`` where it looks something like ``oa_package/08/e0/PMC13900.tar.gz``.
+- You can then download a tar file for a given ``PMID`` or ``PMC`` from ``ftp://ftp.ncbi.nlm.nih.gov/pub/pmc/oa_package/08/e0/PMC13900.tar.gz``. You can check out ``ftp://ftp.ncbi.nlm.nih.gov/pub/pmc/oa_package/`` to get all the access for tar files.
 
[…]
