+++ _sources/install.rst.txt
@@ -16,4 +16,20 @@ or clone the repository and install using ``pip``
     pip install ./pubmed_parser
 
 
-You can test your installation by running ``pytest --cov=pubmed_parser tests/ --verbose``
+You can test your installation by running 
+
+.. code-block:: bash
+
+    pytest --cov=pubmed_parser tests/ --verbose
+
+
+To build a documentation page, change directory to ``docs`` folder and then run
+
+
+.. code-block:: bash
+
+    make html
+
+
+This will build a page in ``_build/html`` which you can open via regular browser.
+You need ``sphinx`` and ``sphinx-gallery`` to build a documentation page.
