+++ _sources/spark.rst.txt
@@ -1,9 +1,12 @@
-Setting up PySpark with Pubmed Parser
+Setting up Pubmed Parser with PySpark
 =====================================
 
 
-We put small snippet to setup Spark 2.1 on Jupyter Notebook here for initializing the notebook. 
-We use Spark workflow to process MEDLINE XML data to Spark dataframe. **Note** that the path to downloaded Spark might be different.
+Below, we put a small snippet to setup ``Spark 2.1`` on Jupyter Notebook. 
+We can use PySpark as a workflow to process MEDLINE XML data to Spark dataframe. 
+Using PySpark can reduce parsing time of more than 25 million documents to less than 10 minutes when you have multiple core processors.
+
+**Note** that the ``spark_home`` path to downloaded Spark might be different.
 
 .. code-block:: python
 
@@ -12,7 +15,7 @@ We use Spark workflow to process MEDLINE XML data to Spark dataframe. **Note** t
     findspark.init(spark_home="/opt/spark-2.1.0-bin-cdh5.9.0/")
 
 
-In Spark 2.1, `spark` in this case can use as `sparkContext` and has access to `parallelize` or `createDataFrame` already.
+In Spark 2.1, ``spark`` in this case can use as ``sparkContext`` which has access to ``parallelize`` or ``createDataFrame``.
 
 .. code-block:: python
 
@@ -37,4 +40,7 @@ In Spark 2.1, `spark` in this case can use as `sparkContext` and has access to `
         config(conf=conf).\
         getOrCreate()
 
-We will update these documentation soon on how to incorporate Pubmed Parser with `dask <https://dask.org/>`_.
+
+Please see the full implementation details in `scripts <https://github.com/titipata/pubmed_parser/tree/master/scripts>`_ folder on the repository.
+
+We will update the documentation on how to incorporate Pubmed Parser with `dask <https://dask.org/>`_ soon.
