+++ docs/doc_requirements.txt
@@ -1,0 +1,3 @@
+sphinx_rtd_theme
+sphinx
+sphinx-gallery
