+++ .travis.yml
@@ -19,6 +19,7 @@ script:
   - pytest --cov=pubmed_parser tests/ --verbose
 
 build_documentation:
+  - pip install -r docs/doc_requirements.txt
   - make html
   - touch _build/html/.nojekyll
 
