### 2020-01-23 eaca1b1
Change directory to docs before make html
