+++ .travis.yml
@@ -28,7 +28,7 @@ deploy:
   keep-history: true
   on:
     branch: master
-  local_dir: _build/html/
+  local_dir: docs/_build/html/
 
 
 after_success:
