+++ .github/ISSUE_TEMPLATE/question.md
@@ -1,10 +1,0 @@
----
-name: Question
-about: Any general question relating to Pubmed Parser
-title: Question
-labels: ''
-assignees: ''
-
----
-
-Please provide a clear description of your question relating to Pubmed Parser e.g. installation, scripts, ...
