### 2020-01-23 bd51e99
update travis.yml, separate steps
