### 2020-01-23 06dbc70
overwrite script and install steps
