+++ .travis.yml
@@ -1,9 +1,5 @@
 language: python
-env:
-  - 3.5
-  - 3.6
-  - 3.7
-  - 3.8
+python: "3.7"
 
 notifications:
   email: false
@@ -20,7 +16,6 @@ jobs:
       script:
         - pytest --cov=pubmed_parser tests/ --verbose
     - stage: "Build Documentation"
-      env: 3.8
       install:
         - pip install -r docs/doc_requirements.txt
       script:
