### 2020-01-23 2f1ac77
install pubmed_parser itself too
