+++ .travis.yml
@@ -7,12 +7,11 @@ notifications:
 before_install:
   - python --version
   - pip install -U pip
+  - python setup.py install
 
 jobs:
   include:
     - stage: test
-      install:
-        - python setup.py install
       script:
         - pytest --cov=pubmed_parser tests/ --verbose
     - stage: "Build Documentation"
