+++ .travis.yml
@@ -20,15 +20,15 @@ jobs:
       script:
         - cd docs && make html
         - touch _build/html/.nojekyll
-    - stage: "Deploy Documentation"
-      deploy:
-        provider: pages
-        skip_cleanup: true
-        github_token: $GITHUB_TOKEN  # Set in travis-ci.org dashboard, marked secure
-        keep-history: true
-        on:
-          branch: master
-        local_dir: docs/_build/html/
+
+deploy:
+  provider: pages
+  skip_cleanup: true
+  github_token: $GITHUB_TOKEN
+  keep-history: true
+  on:
+    branch: master
+  local_dir: docs/_build/html/
 
 after_success:
   - bash <(curl -s https://codecov.io/bash)
