### 2020-01-23 75b87da
move deploy into documentation step, run codecov after test success
