+++ .travis.yml
@@ -14,21 +14,20 @@ jobs:
     - stage: test
       script:
         - pytest --cov=pubmed_parser tests/ --verbose
-    - stage: "Build Documentation"
+      after_success:
+        - bash <(curl -s https://codecov.io/bash)
+    - stage: "Documentation"
       install:
         - pip install -r docs/doc_requirements.txt
       script:
         - cd docs && make html
         - touch _build/html/.nojekyll
+      deploy:
+        provider: pages
+        skip_cleanup: true
+        github_token: $GITHUB_TOKEN
+        keep-history: true
+        on:
+          branch: master
+        local_dir: docs/_build/html/
 
-deploy:
-  provider: pages
-  skip_cleanup: true
-  github_token: $GITHUB_TOKEN
-  keep-history: true
-  on:
-    branch: master
-  local_dir: docs/_build/html/
-
-after_success:
-  - bash <(curl -s https://codecov.io/bash)
