### 2020-01-23 cbb5d3e
add env to test stage
