+++ .travis.yml
@@ -1,4 +1,5 @@
 language: python
+python: "3.7"
 
 notifications:
   email: false
@@ -11,11 +12,6 @@ before_install:
 jobs:
   include:
     - stage: test
-      env:
-        - 3.5
-        - 3.6
-        - 3.7
-        - 3.8
       script:
         - pytest --cov=pubmed_parser tests/ --verbose
       after_success:
