### 2020-01-24 7c95063
Handle field.text being None
pmid.text being None and citation.text being none broke pmid.text.strip() and citation.text.strip(), respectively.

This happened in medline update file 1051.
