### 2020-01-27 titipata/pubmed_parser#82 @tleonardi
Missing details in API docs
Hi,
I've noticed some functions miss the sphinx documentation:
parse_pubmed_paragraph(): the parameter all_paragraph is not documented
table_to_df(): parameters and return value not documented
parse_pubmed_web_tree(): parameters and return value not documented
extract_citations():  parameters and return value not documented
extract_pmc():  parameters and return value not documented
