### 2020-01-27 #1979 @tleonardi
Hi @titipata,
I'm happy with all the changes you've made, great job writing and automating the tests!
I only have a few concerns left, mostly regarding the README and documentation.

I think it would be important, at the beginning of the README, to describe what kind of information is contained in the XML files and what can be obtained with pubmed_parser.
You could also add an index with links to the various sections of the README (Installation, Documentation, Examples, Citation, etc...).
Regarding the documentation, I think many functions lack a high level explanation of what they do: what do the input XMLs contain? What's contained in the results returned?
Lastly, I've just opened an [issue](https://github.com/titipata/pubmed_parser/issues/82) indicating some missing bits in the API docs.
All the rest is looking good!
