### 2020-01-27 #1979 @titipata
@tleonardi Thank you so much for the details comments! I'll update the API documentation according to your comment by the end of this week. I will also edit the README so that it contains information about the XML files. I'll tag you once it's done.
