### 2020-01-28 7d4ae40
Update API docs according to #82
