### 2020-01-29 #1979 @timClicks
@titipata Thank you for taking the time to publish this package as open source software. I've reviewed it and the paper, and have nothing further to add above @tleonardi's comments.

@majensen I have reviewed the submission and recommend accepting it for publication.
