+++ pubmed_parser/medline_parser.py
@@ -32,7 +32,13 @@ def parse_pmid(pubmed_article):
         article_ids = pubmed_article.find("PubmedData/ArticleIdList")
         if article_ids is not None:
             pmid = article_ids.find('ArticleId[@IdType="pmid"]')
-            pmid = pmid.text.strip() if pmid is not None and pmid.text is not None  else ""
+            if pmid is not None:
+                if pmid.text is not None:
+                    pmid = pmid.text.strip()
+                else:
+                    pmid = ""
+            else:
+                pmid = ""
         else:
             pmid = ""
     return pmid
@@ -435,7 +441,13 @@ def parse_references(pubmed_article, reference_list):
     if reference_list_data is not None:
         for ref in reference_list_data.findall("Reference"):
             citation = ref.find("Citation")
-            citation = citation.text.strip() if citation is not None and citation.text is not None else ""
+            if citation is not None:
+                if citation.text is not None:
+                    citation = citation.text.strip()
+                else:
+                    citation = ""
+            else:
+                citation = ""
             article_ids = ref.find("ArticleIdList")
             pmid = (
                 article_ids.find('ArticleId[@IdType="pubmed"]')
