### 2020-02-01 titipata/pubmed_parser#82 @titipata
I've fixed this issue in 7d4ae40. I will further update the documentation and point to this issue later on!
