### 2020-02-01 67a0df1
(ed.) - tighten up some phrasing
