### 2020-02-01 #1979 @majensen
Hi @titipata - I've created a PR for you to consider: https://github.com/titipata/pubmed_parser/pull/84

Also, in the code example in the text, would you consider changing
`data/medline16n0902.xml.gz`
to
`data/pubmed20n0014.xml.gz`
since the latter file is actually present in the `data` directory of the repo? Then a user can simply run the sample code after cloning and installing.

thanks
