### 2020-02-01 acbf58d
Update MEDLINE file in paper.md
