### 2020-02-01 8ebc47a
I missed one of the pmid.text.strips() in my last fix. Here's the fix for the one I missed using your new styling.
