### 2020-02-03 #1979 @tleonardi
Hi @majensen,
I just hadn't checked the last changes that @titipata introduced after my last comments. I have checked now and the API docs look good! I haven't seen changes in the README, but my suggestions where mostly minor details. No need to hold the publication because of this.
Overall, `pubmed_parser` is a solid software that's going to be of great help to the scientific community (as shown by the number of publications that already use it). I am happy to recommend it for publication in JOSS.
