### 2020-02-03 #1979 @majensen
@titipata we are on the home stretch. Can you now create a persistent archive of the repository at a service such as [Zenodo](https://zenodo.org)? For Zenodo there is a mini tutorial at [this comment](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729).
Thanks!
