### 2020-02-06 29ad825
Update version to 0.2.2
