### 2020-02-06 #1979 @titipata
Hi @majensen, thank you so much. I updated DOI for the repository and updated it in `README` accordingly. The new DOI should be `10.5281/zenodo.3648715` for the Pubmed Parser 0.2.2
