### 2020-02-07 #1979 @majensen
Thanks @titipata  - JOSS requires the author list in the paper to be the same as the author list in the archive. I think it's great to give credit to your contributors, but I think GitHub handles that pretty well, so I would suggest updating the archive metadata author list to include just the 3 on your initial submission.
