### 2020-02-07 #1979 @titipata
@majensen thank you so much! I updated the metadata author list to 3, the same as in the submission in Pubmed Parser 0.2.2.
