### 2020-02-07 #1979 @danielskatz
The other option would be to add the others to the paper as well.  We just want these to be consistent.
