### 2020-02-07 #1979 @titipata
@danielskatz thanks! We have discussed and decided to include 3 main contributors in our paper. I edited Zenodo accordingly. This should be consistent with the JOSS submission :) Let me know if there is any concern regarding our decision!
