+++ paper/paper.bib
@@ -73,18 +73,17 @@
 }
 
 @inproceedings{mesbah2018smartpub,
-  title={SmartPub: A Platform for Long-Tail Entity Extraction from Scientific Publications},
+  title={{SmartPub}: a platform for long-tail entity extraction from scientific publications},
   author={Mesbah, Sepideh and Bozzon, Alessandro and Lofi, Christoph and Houben, Geert-Jan},
-  booktitle={WWW},
+  booktitle={Companion Proceedings of the {The Web Conference} 2018},
   pages={191--194},
-  year={2018},
-  organization={International World Wide Web Conferences Steering Committee}
+  year={2018}
 }
 
 @inproceedings{tang2019parallel,
-  title={Parallel Computing for Large-Scale Author Name Disambiguation in MEDLINE},
+  title={Parallel Computing for Large-Scale Author Name Disambiguation in {MEDLINE}},
   author={Tang, Anyao and Wu, Chengkun and Liu, Jie and Wang, Wei and Yang, Xi and Xing, Yuting},
-  booktitle={2019 IEEE 21st International Conference on High Performance Computing and Communications; IEEE 17th International Conference on Smart City; IEEE 5th International Conference on Data Science and Systems (HPCC/SmartCity/DSS)},
+  booktitle={2019 {IEEE} 21st {International Conference on High Performance Computing and Communications}; {IEEE} 17th {International Conference on Smart City}; {IEEE} 5th {International Conference on Data Science and Systems (HPCC/SmartCity/DSS)}},
   pages={1580--1586},
   year={2019},
   organization={IEEE},
@@ -92,7 +91,7 @@
 }
 
 @article{abdeddaim2018mesh,
-  title={The MeSH-gram Neural Network Model: Extending Word Embedding Vectors with MeSH Concepts for UMLS Semantic Similarity and Relatedness in the Biomedical Domain},
+  title={The {MeSH-gram Neural Network Model}: Extending Word Embedding Vectors with {MeSH} Concepts for {UMLS} Semantic Similarity and Relatedness in the Biomedical Domain},
   author={Abdedda{\"\i}m, Sa{\"\i}d and Vimard, Sylvestre and Soualmia, Lina Fatima},
   journal={arXiv preprint arXiv:1812.02309},
   url={https://arxiv.org/abs/1812.02309},      
