### 2020-02-07 #1979 @kthyng
@titipata Your paper looked good but your references had typos. I fixed them in a PR. See if you want to merge the changes: [#86](https://github.com/titipata/pubmed_parser/pull/86)
