### 2020-02-08 #1979 @majensen
@titipata I believe you can update the archive with a new version and keep the DOI- just report back the new version here. Thanks!
