### 2020-02-08 a4bcf4e
Update DOI for JOSS
