### 2020-02-08 #1979 @kthyng
Congratulations to  @titipata on your new publication! Thanks to @majensen for editing, and to @timClicks and @tleonardi for your time and expertise reviewing.
